+++ controllers/api/donorDash.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { Donor, Appointment, } = require('../../models');
+const { Donor, Appointment } = require('../../models');
 const withAuth = require('../../utils/auth');
 const { Op } = require("sequelize");
 
@@ -12,61 +12,68 @@ router.get('/', async (req, res) => {
     } catch (err) {
       res.status(500).json(err);
     }
-  });
+});
 
-  //Get donor data by id Insomnia root: http://localhost:3001/api/donorDash
-  /*Eaxample of JSON object for testing with Insomnia
-  {
-    "email": taz@hotmail.com
-    "password": "apple101",
-  } */
-  //TO DO: join with the Appointment table and find next appointment return all donor data (excluding password) and  next date As nextdate_don and time as nexttime_don 
+//Get donor data by id Insomnia root: http://localhost:3001/api/donorDash
+/*Eaxample of JSON object for testing with Insomnia
+{
+  "email": taz@hotmail.com
+  "password": "apple101",
+} */
+//TO DO: join with the Appointment table and find next appointment return all donor data (excluding password) and  next date As nextdate_don and time as nexttime_don 
     
-  //[1] return a donor's details  [WORKING]
-    router.get('/data', withAuth, async (req, res) => {
-    try {
-      const donorData = await Donor.findByPk(req.session.user_id, {
-        attributes: { exclude: ['password'] }
-
-
+//[1] return a donor's details  [WORKING]
+  router.get('/data', withAuth, async (req, res) => {
+  try {
+    const donorData = await Donor.findByPk(req.session.user_id, {
+      attributes: { exclude: ['password'] }
       });
+
+      const donor = donorData.get({ plain: true})
+        
+      res.render('donor', { donor })   
       res.status(200).json(donorData);
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+
+
+
+/*[2] Donor's Next appointment: using session.user_id to identify the donor.  Currently set to return all appointments between current date and 16 weeks in the future.
+//This range was chosen to capture the minimum allowed time between donation appointments, which is 12 weeks for men and 16 for women WORKING*/
+
+router.get('/nextappointment', withAuth, async (req, res) => {
+  
+  console.log("next app")
+  try {
+      const startDate = new Date();
+      const endDateTemp = new Date();
+      endDateTemp.setDate(endDateTemp.getDate() + 112)
+      const endDate = new Date(endDateTemp);
+
+      console.log(startDate, endDate)
+      console.log(startDate.toLocaleDateString(), endDate.toLocaleDateString())
+
+        const donorData = await Appointment.findByPk(req.session.user_id,{ 
+          where: {
+            date: {
+              [Op.between]: [startDate, endDate]
+            }
+          },
+        
+        });
+
+        const donors = donorData.map((donor) =>
+        donor.get({ plain: true})
+        );
+        res.render('donor', { donors })
+        res.status(200).json(donorData);
     } catch (err) {
       res.status(500).json(err);
     }
-  });
-
-
-
-
-  /*[2] Donor's Next appointment: using session.user_id to identify the donor.  Currently set to return all appointments between current date and 16 weeks in the future.
-  //This range was chosen to capture the minimum allowed time between donation appointments, which is 12 weeks for men and 16 for women WORKING*/
-
-  router.get('/nextappointment', withAuth, async (req, res) => {
-    
-    console.log("next app")
-    try {
-        const startDate = new Date();
-        const endDateTemp = new Date();
-        endDateTemp.setDate(endDateTemp.getDate() + 112)
-        const endDate = new Date(endDateTemp);
-
-        console.log(startDate, endDate)
-        console.log(startDate.toLocaleDateString(), endDate.toLocaleDateString())
-
-         const donorData = await Appointment.findByPk(req.session.user_id,{ 
-            where: {
-              date: {
-                [Op.between]: [startDate, endDate]
-              }
-            },
-          
-          });
-      res.status(200).json(donorData);
-    } catch (err) {
-      res.status(500).json(err);
-    }
-  });
+});
 
 
 
@@ -74,11 +81,11 @@ router.get('/', async (req, res) => {
 
  
 
-  /*[3] Donor appointment history: using session.user_id to identify the donor.  Returns the donors appointments between current date and 365 days in the past.  
-  This could be changed to allow donor to select the range by passing a value in the request body. WORKING*/
-  router.get('/appointmentHist', withAuth, async (req, res) => {
-       
-    try {
+/*[3] Donor appointment history: using session.user_id to identify the donor.  Returns the donors appointments between current date and 365 days in the past.  
+This could be changed to allow donor to select the range by passing a value in the request body. WORKING*/
+router.get('/appointmentHist', withAuth, async (req, res) => {
+      
+  try {
         const startDate = new Date();
         const endDateTemp = new Date();
         endDateTemp.setDate(endDateTemp.getDate() - 365)
@@ -87,14 +94,20 @@ router.get('/', async (req, res) => {
         console.log(startDate, endDate)
         console.log(startDate.toLocaleDateString(), endDate.toLocaleDateString())
 
-         const donorData = await Appointment.findByPk(req.session.user_id,{
+          const donorData = await Appointment.findByPk(req.session.user_id,{
             where: {
               date: {
                 [Op.between]: [startDate, endDate]
               }
             },
           });
-      res.status(200).json(donorData);
+        
+        const donors = donorData.map((donor) =>
+        donor.get({ plain: true})
+        );
+        res.render('donor', { donors })
+        res.status(200).json(donorData);
+
     } catch (err) {
       res.status(500).json(err);
     }
@@ -111,7 +124,11 @@ router.get('/', async (req, res) => {
         }
                 
       });          
-       
+      const donors = donorData.map((donor) =>
+      donor.get({ plain: true})
+      );
+
+      res.render('donor', { donors }) 
       res.status(200).json(donorData);
     } catch (err) {
       res.status(500).json(err);
@@ -119,26 +136,31 @@ router.get('/', async (req, res) => {
   });
 
 //[5] donor's next appointment - this needs more seeds to test it's working properly NEEDS MORE TESTING
-  router.get('/appointmentCurr', withAuth, async (req, res) => {
-    console.log("next app")
-    try {
-        const startDate = new Date();
-        console.log(startDate);
-        console.log(startDate.toLocaleDateString());
-       
-         const donorData = await Appointment.findAll({          
-            where: {
-              date: startDate
-            },
-          
-          });
-      res.status(200).json(donorData);
-    } catch (err) {
-      res.status(500).json(err);
-    }
-  });
+router.get('/appointmentCurr', withAuth, async (req, res) => {
+  console.log("next app")
+  try {
+      const startDate = new Date();
+      console.log(startDate);
+      console.log(startDate.toLocaleDateString());
+      
+        const donorData = await Appointment.findAll({          
+          where: {
+            date: startDate
+          },
+        
+        });
+      const donors = donorData.map((donor) =>
+      donor.get({ plain: true})
+      );
 
-  // [6]Delete donor account - probably should only be done from from Admin side? WORKING
+      res.render('donor', { donors }) 
+      res.status(200).json(donorData);
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+// [6]Delete donor account - probably should only be done from from Admin side? WORKING
 
 router.delete('/remove', withAuth, async (req, res) => {
   try {
