+++ public/js/register.js
@@ -11,9 +11,9 @@ const postNewDonor = async (newDonor) => {
       });
 
       if (response.ok) {
-        //the administrator login process was succesful
+        //the donor login process was succesful
         console.log(response.id);
-        document.location.replace('/admin');
+        document.location.replace('/donor');
         alert(response.statusText);
       }
       else {
