+++ views/admin2.handlebars
@@ -15,6 +15,7 @@
         <li class="nav-item">
           <p class="nav-link active">Search for donor</p>
           <div class="input-group mb-3">
+            <a class="btn btn-primary" href="#" role="button">Link</a>
             <input type="text" class="form-control" placeholder="Recipient's username" aria-label="Recipient's username" aria-describedby="button-addon2">
             <button class="btn btn-outline-secondary" type="button" id="button-addon2">Search</button>
           </div>
