+++ views/donorOld.handlebars
@@ -1,0 +1,35 @@
+
+
+
+
+    <header>      
+      <h1> Donor Dashboard </h1>
+      <h3>Welcome {{first_name}} {{last_name}}</h3>
+      <div class="d-grid  d-md-flex justify-content-md-end">
+        <button class="btn btn-primary me-md-2" id="logout" type="button">Logout</button>
+        </div>
+    </header>
+
+    {{!-- Put name and date of birth in and change cards  --}}
+
+    <ul class="nav flex-column" style="margin-top: 10px; padding-top: 75px">
+      <li class="nav-item">
+        <a class="nav-link active" href="#">View My Details
+      </a>
+      <li class="nav-item">
+          <a class="nav-link active" href="#">Update Details
+        </a>
+        <li class="nav-item">
+          <a class="nav-link active" href="#">Appointment History
+        </a>
+        <li class="nav-item">
+          <a class="nav-link active" href="#">View Next Appointment
+        </a>
+        <li class="nav-item">
+          <a class="nav-link active" href="#">Blood Donation Record
+        </a>
+        <li class="nav-item">
+          <a class="nav-link active" href="#">View Blood Type
+        </a></li></ul>
+
+<script src="./js/donorDash.js"></script>
