+++ views/login.handlebars
@@ -6,7 +6,7 @@
               
               <ul class="navbar-nav">
                 <li class="nav-item">
-                  <a class="nav-link active" aria-current="page" href="#">Home</a>
+                   <div> <a  class="btn btn-danger"  href="/"> Home</a></div>
                 </li>
             </ul>
             </div>
